requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
blobfile==2.1.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.6.0
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.21.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.7.0
email_validator==2.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpy==0.3.3
filelock==3.13.1
Flask==3.0.3
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.39.0
gradio_client==1.1.1
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.6
imageio==2.34.2
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.21.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonmerge==1.9.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
-e git+https://github.com/crowsonkb/k-diffusion.git@ab527a9a6d347f364e3d185ba6d714e22d80cb3c#egg=k_diffusion
kiwisolver==1.4.5
kneed==0.8.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.6
lpips==0.1.4
lxml==4.9.4
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.28.0
opencv-python==4.9.0.80
optuna==3.6.1
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-lightning==1.4.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
resize-right==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
ruff==0.5.5
safetensors==0.4.2
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.11.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.17.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tifffile==2024.7.24
timm==1.0.11
tinycss2==1.4.0
tokenizers==0.15.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.2.0
torchdiffeq==0.2.4
torchmetrics==0.6.0
torchsde==0.2.6
torchvision==0.17.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
trampoline==0.1.2
transformers==4.37.2
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.30.3
uvloop==0.19.0
wandb==0.17.5
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
